chore(Button): migrate javascript to typescript
